WIP: Writing out bad ppm.

+++ options.go
@@ -2,8 +2,8 @@ package main
 
 // Options represents rendering configuration.
 type Options struct {
-	width         int64
-	height        int64
+	width         int
+	height        int
 	fov           float64
 	cameraToWorld Matrix44f
 }
